rename ChemPage->ResultsPage for clarity
