+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import './App.css';
 import SiteHeader from "./components/SiteHeader";
 import SiteFooter from './components/SiteFooter.tsx';
-import ChemPage, { MoleculeInfo } from "./components/ChemPage.tsx";
+import ResultsPage, { MoleculeInfo } from "./components/ResultsPage.tsx";
 import SearchResults from "./components/form/SearchResults.tsx";
 import { createTheme, ThemeProvider } from "@mui/material";
 
@@ -15,7 +15,7 @@ function App() {
                 <div className="page-container">
                     <SiteHeader />
                     {!!chem ? (
-                        <ChemPage result={chem.result} setChem={setChem} canGetDrugInfo={chem.canGetDrugInfo} canGetTargetInfo={chem.canGetTargetInfo}/>
+                        <ResultsPage result={chem.result} setChem={setChem} canGetDrugInfo={chem.canGetDrugInfo} canGetTargetInfo={chem.canGetTargetInfo}/>
                     ) : (
                         <>
                             <SearchResults setChem={setChem} />
